Partfix 4092 (#4144)
* CM test workflow

* fix pep 8

* update definitions.lic file

* add icons

* update default ratios for age, gender and vulnerability concepts
add sources and tweak descriptions

* tweak values to add to 1

* fix svg

* add infant svg

* concept definition updates

* concept definition updates

* concept definition updates

* fix failing tests

* fix failing tests

* make Travis and Sunni happy

+++ safe/impact_function/test/test_impact_function.py
@@ -485,7 +485,7 @@ class TestImpactFunction(unittest.TestCase):
             displaced_field['field_name'])
         displaced_population = analysis.uniqueValues(index)[0]
         self.assertEqual(
-            displaced_population * female_ratio, female_displaced)
+            int(displaced_population * female_ratio), female_displaced)
 
         # Check that we have more than 0 youth displaced in the analysis layer
         index = analysis.fieldNameIndex(
